task2_animal_ner_vision/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
datasets>=2.17.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
Pillow>=10.0.0
jupyter>=1.0.0
